+++ backend/muster-roll/src/main/java/org/egov/repository/querybuilder/MusterRollQueryBuilder.java
@@ -10,6 +10,8 @@ import org.springframework.stereotype.Component;
 import java.util.Collection;
 import java.util.List;
 
+import static org.egov.common.utils.MultiStateInstanceUtil.SCHEMA_REPLACE_STRING;
+
 @Component
 public class MusterRollQueryBuilder {
 
@@ -19,16 +21,16 @@ public class MusterRollQueryBuilder {
     private static final String FETCH_MUSTER_ROLL_QUERY = "SELECT muster.id,muster.tenant_id,muster.musterroll_number,muster.attendance_register_id,muster.status,muster.musterroll_status,muster.start_date,muster.end_date,muster.createdby,muster.lastmodifiedby,muster.createdtime,muster.lastmodifiedtime,muster.additionaldetails,muster.reference_id,muster.service_code,"+
             "ind.id AS summaryId,ind.muster_roll_id AS indMusterId,ind.individual_id AS IndividualId,ind.actual_total_attendance AS actualTotalAttendance,ind.additionaldetails AS indAddlDetails,ind.createdby AS indCreatedBy,ind.lastmodifiedby AS indModifiedBy,ind.createdtime AS indCreatedTime,ind.lastmodifiedtime AS indModifiedTime,ind.modified_total_attendance AS modifiedTotalAttendance,"+
             "attn.id AS attendanceId,attn.attendance_summary_id AS attnSummaryId,attn.date_of_attendance AS AttnDate,attn.attendance_value AS attendance,attn.additionaldetails AS attnAddlDetails,attn.createdby AS attnCreatedBy,attn.lastmodifiedby AS attnModifiedBy,attn.createdtime AS attnCreatedTime,attn.lastmodifiedtime AS attnModifiedTime "+
-            "FROM eg_wms_muster_roll AS muster " +
+            "FROM " + SCHEMA_REPLACE_STRING + ".eg_wms_muster_roll AS muster " +
             "LEFT JOIN " +
-            "eg_wms_attendance_summary AS ind " +
+            SCHEMA_REPLACE_STRING + ".eg_wms_attendance_summary AS ind " +
             "ON (muster.id=ind.muster_roll_id) " +
             "LEFT JOIN " +
-            "eg_wms_attendance_entries AS attn " +
+            SCHEMA_REPLACE_STRING + ".eg_wms_attendance_entries AS attn " +
             "ON (attn.attendance_summary_id=ind.id) ";
 
     private static final String MUSTER_ROLL_COUNT_QUERY = "SELECT distinct(muster.id) " +
-            "FROM eg_wms_muster_roll AS muster ";
+            "FROM  "+ SCHEMA_REPLACE_STRING + ".eg_wms_muster_roll AS muster ";
 
     private static final String PAGINATION_WRAPPER = "SELECT * FROM " +
             "(SELECT *, DENSE_RANK() OVER (ORDER BY {sortBy} {orderBy}) offset_ FROM " +